:whale: docker dev fix, :heavy-minus-sign: removed unused dependencies
