+++ api/package.json
@@ -11,11 +11,11 @@
         "serve-debug": "nodemon --inspect dist/server.js",
         "start": "npm run serve",
         "test": "jest -i",
-        "watch": "concurrently -k -p \"[{name}]\" -n \"TypeScript,Node\" -c \"cyan.bold,green.bold\" \"npm run watch-ts\" \"npm run watch-node\"",
+        "watch": "npm run build && concurrently -k -p \"[{name}]\" -n \"TypeScript,Node\" -c \"cyan.bold,green.bold\" \"npm run watch-ts\" \"npm run watch-node\"",
         "watch-debug": "concurrently -k -p \"[{name}]\" -n \"TypeScript,Node\" -c \"yellow.bold,green.bold\" \"npm run watch-ts\" \"npm run serve-debug\"",
         "watch-node": "nodemon dist/server.js",
         "watch-test": "npm run test -- --watchAll",
-        "watch-ts": "tsc -w"
+        "watch-ts": "tsc -w --preserveWatchOutput"
     },
     "jest": {
         "collectCoverage": true,
@@ -35,7 +35,6 @@
         "axios-cookiejar-support": "^0.5.1",
         "bcrypt": "^4.0.1",
         "bcrypt-nodejs": "^0.0.3",
-        "body-parser": "^1.19.0",
         "compression": "^1.7.4",
         "connect-redis": "^4.0.4",
         "cors": "^2.8.5",
@@ -64,11 +63,9 @@
         "@types/async": "^3.2.0",
         "@types/bcrypt": "^3.0.0",
         "@types/bcrypt-nodejs": "^0.0.31",
-        "@types/body-parser": "^1.19.0",
         "@types/chai": "^4.2.11",
         "@types/compression": "^1.7.0",
         "@types/concurrently": "^5.1.0",
-        "@types/connect-mongo": "^3.1.3",
         "@types/cors": "^2.8.6",
         "@types/errorhandler": "^1.5.0",
         "@types/eslint": "^6.1.9",
@@ -83,7 +80,6 @@
         "@types/node": "^13.11.0",
         "@types/nodemailer": "^6.4.0",
         "@types/passport": "^1.0.3",
-        "@types/passport-facebook": "^2.1.9",
         "@types/passport-local": "^1.0.33",
         "@types/request": "^2.48.4",
         "@types/request-promise": "^4.1.46",
