+++ client/Dockerfile
@@ -2,22 +2,19 @@ FROM node:12-buster AS base
 ENV PATH=$PATH:/usr/src/app/node_modules/.bin \
     NODE_ENV=development \
     CI=true
+RUN apt-get update && \ 
+    apt-get -y install \
+    libxi6 \
+    libgl1-mesa-glx
 WORKDIR /usr/src/app
 COPY package*.json ./
 RUN npm install
 
-FROM base as deps
-RUN apt-get update && \ 
-    apt-get -y install \
-    libglu1 \
-    libxi6 \
-    libgconf-2-4
-
-FROM deps as dev
+FROM base as dev
 EXPOSE 3000
 CMD npm run start
 
-FROM deps as build
+FROM base as build
 COPY . .
 RUN npm run build --if-present
